+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## v0.3.2 - Model Package 
   * Commit: 
   * Model fields
-  * Schema type 
+  * Schema type
+  * Clock format tests
 ## v0.3.1 - My Package 
   * Commit: 2026-03-31 15:26
   * Request type